refactor: 학생 다운로드 util 함수를 shared/utils로 이동
