--- apps/graduate/src/shared/components/StudentDownload/index.ts
@@ -1,3 +1,0 @@
-export { downloadStudentDetailExcel } from './downloadStudentDetailExcel';
-export { downloadStudentDetailFromRows } from './downloadStudentDetailFromRows';
-export { handleDownload } from './handleDownload';
