+++ apps/graduate/src/shared/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './studentDownload';
